+++ code/system/sys.util/src/Value.Object/index.ts
@@ -1,0 +1,2 @@
+export * from './Value.Object.path';
+export * from './Value.Object';
