+++ code/system/sys.util/src/Value/Value.ts
@@ -1,8 +1,9 @@
 import { Is } from '../Is';
+import * as Object from '../Value.Object';
+
 import * as Array from './Value.Array';
 import * as Hash from './Value.Hash';
 import * as Math from './Value.Math';
-import * as Object from './Value.Object';
 import * as Random from './Value.Random';
 import * as To from './Value.To';
 import * as Util from './Value.u';
